refactor(sessions): split launch admission out of the config handle
`handle.rs` grew past the 600-line repo cap (PROD-SIZE-1) when the absent-
control disposition and its tests landed. Split along the real seam rather
than record a net-new ratchet exception.

`admission.rs` now owns the pure decisions the start path makes BEFORE
anything is sent to the agent: whether a requested model id resolves to one
the live session offers, and what to do with a requested control the live
statement does not offer under the applied model. `handle.rs` keeps the async
apply orchestration and re-exports the seam, so every existing
`config::handle::…` path and the actor test module resolve unchanged.

No behavior change.
