+++ anyharness/crates/anyharness-lib/src/live/sessions/actor/config/handle.rs
@@ -1,11 +1,8 @@
 use agent_client_protocol as acp;
-use anyharness_contract::v1::{
-    ConfigApplyState, RawSessionConfigOption, SessionConfigOptionType,
-};
+use anyharness_contract::v1::ConfigApplyState;
 
 use crate::domains::sessions::launch_intent::ResolvedLaunchIntent;
-use crate::domains::sessions::live_config::controls::option_matches_key;
-use crate::domains::sessions::live_config::{NormalizedControlKind, LEGACY_MODE_COMPAT_CONFIG_ID};
+use crate::domains::sessions::live_config::LEGACY_MODE_COMPAT_CONFIG_ID;
 use crate::live::sessions::actor::config::apply::{
     apply_mode_via_direct_setter_legacy, apply_specific_config_option, try_apply_model_preference,
 };
@@ -13,13 +10,17 @@ use crate::live::sessions::actor::config::confirmation::config_value_matches_cur
 use crate::live::sessions::actor::config::persist::persist_requested_config_value_if_changed;
 use crate::live::sessions::actor::config::queue::queue_pending_config_change;
 use crate::live::sessions::actor::config::selection::{
-    find_select_option_by_purpose, find_select_option_for_request, into_raw_pending_option,
-    select_option_values,
+    find_select_option_by_purpose, find_select_option_for_request, select_option_values,
 };
 use crate::live::sessions::actor::config::types::{
     tracked_config_purpose, ConfigApplyOutcome, ConfigPurpose,
 };
 use crate::live::sessions::actor::state::{SessionActor, SessionStartupState};
+
+pub(in crate::live::sessions::actor) use super::admission::{
+    absent_control_is_posture, initial_control_disposition, resolve_requested_model_id,
+    InitialControlDisposition,
+};
 
 /// Applies the immutable create-time intent and accepts only a value that the
 /// live harness statement both advertises and positively confirms. Model is
@@ -67,7 +68,12 @@ pub(in crate::live::sessions::actor) async fn apply_resolved_launch_intent(
                     Some(resolved)
                 }
                 None => {
-                    log_initial_config_apply(session_id, agent_kind, "model", "membership_rejected");
+                    log_initial_config_apply(
+                        session_id,
+                        agent_kind,
+                        "model",
+                        "membership_rejected",
+                    );
                     anyhow::bail!(
                         "requested model '{model_id}' is absent from the live {agent_kind} session"
                     );
@@ -98,7 +104,10 @@ pub(in crate::live::sessions::actor) async fn apply_resolved_launch_intent(
             confirmed_result_code(outcome),
         );
         anyhow::ensure!(
-            matches!(outcome, ConfigApplyOutcome::NoChange | ConfigApplyOutcome::AppliedAuthoritative),
+            matches!(
+                outcome,
+                ConfigApplyOutcome::NoChange | ConfigApplyOutcome::AppliedAuthoritative
+            ),
             "requested model '{model_id}' was not confirmed by the live {agent_kind} session"
         );
     }
@@ -249,86 +258,6 @@ pub(in crate::live::sessions::actor) async fn apply_resolved_launch_intent(
     Ok(())
 }
 
-/// How the start path treats one explicit control value against the live
-/// statement. Create-time validation runs against the harness-level
-/// observation, but some harnesses narrow a QUALITY control's value set per
-/// model (codex reasoning_effort). A quality value the live session does not
-/// OFFER after the model applied is dropped to the session default rather than
-/// failing the whole start. The membership check runs before anything is sent;
-/// an OFFERED value whose setter read-back refuses it stays fatal — the same
-/// invariant the legacy mode branch keeps.
-///
-/// Posture controls are never dropped: launching a collaboration mode, mode /
-/// approval policy or sandbox mode at the harness default after the user
-/// explicitly selected against it is a silent behavior change, which is
-/// strictly worse than refusing the start.
-#[derive(Debug, PartialEq, Eq)]
-pub(in crate::live::sessions::actor) enum InitialControlDisposition {
-    AlreadyLive,
-    Apply,
-    Drop,
-    Refuse,
-}
-
-pub(in crate::live::sessions::actor) fn initial_control_disposition(
-    startup_state: &SessionStartupState,
-    config_id: &str,
-    value: &str,
-) -> InitialControlDisposition {
-    if config_value_matches_current_state(startup_state, config_id, value) {
-        return InitialControlDisposition::AlreadyLive;
-    }
-    let offered = find_select_option_for_request(&startup_state.config_options, config_id)
-        .is_some_and(|option| {
-            select_option_values(option).iter().any(|candidate| candidate == value)
-        });
-    if offered {
-        InitialControlDisposition::Apply
-    } else if is_posture_control(startup_state, config_id) {
-        InitialControlDisposition::Refuse
-    } else {
-        InitialControlDisposition::Drop
-    }
-}
-
-/// A posture control decides what the agent is allowed to DO — collaboration
-/// mode, the mode / approval-policy family, sandbox mode. Only quality and
-/// model-narrowing controls are eligible for the soft drop.
-fn is_posture_control(startup_state: &SessionStartupState, config_id: &str) -> bool {
-    if config_id == LEGACY_MODE_COMPAT_CONFIG_ID {
-        return true;
-    }
-    find_select_option_for_request(&startup_state.config_options, config_id).is_some_and(|option| {
-        let raw = into_raw_pending_option(option);
-        option_matches_key(&raw, NormalizedControlKind::Mode)
-            || option_matches_key(&raw, NormalizedControlKind::CollaborationMode)
-    })
-}
-
-/// Whether a control id the live statement never surfaced is a posture control.
-///
-/// `is_posture_control` classifies a live option, so it cannot answer for an id
-/// that is absent: the lookup returns `None` and every absent id would read as
-/// non-posture. Classifying the bare id through the same `option_matches_key`
-/// vocabulary keeps one definition of posture instead of a second id list that
-/// could drift from it.
-fn absent_control_is_posture(config_id: &str) -> bool {
-    if config_id == LEGACY_MODE_COMPAT_CONFIG_ID {
-        return true;
-    }
-    let probe = RawSessionConfigOption {
-        id: config_id.to_string(),
-        name: String::new(),
-        description: None,
-        category: None,
-        option_type: SessionConfigOptionType::Select,
-        current_value: String::new(),
-        options: Vec::new(),
-    };
-    option_matches_key(&probe, NormalizedControlKind::Mode)
-        || option_matches_key(&probe, NormalizedControlKind::CollaborationMode)
-}
-
 pub(in crate::live::sessions::actor) fn intent_without_dropped_controls(
     intent: &ResolvedLaunchIntent,
     dropped_control_ids: &[String],
@@ -397,7 +326,10 @@ async fn try_apply_config_option_by_id(
     else {
         return Ok(ConfigApplyOutcome::NotApplied);
     };
-    if !select_option_values(option).iter().any(|candidate| candidate == value) {
+    if !select_option_values(option)
+        .iter()
+        .any(|candidate| candidate == value)
+    {
         return Ok(ConfigApplyOutcome::NotApplied);
     }
     crate::live::sessions::actor::config::apply::apply_select_config_option(
@@ -408,44 +340,6 @@ async fn try_apply_config_option_by_id(
         value,
     )
     .await
-}
-
-/// Resolves the create-time model id against the ids the live session offers.
-///
-/// Exact membership stays the primary rule. When the exact id is absent, a
-/// live id whose bracket-stripped base equals the requested id's base is the
-/// same model selection under a rotated context-variant id, and resolution
-/// succeeds only when exactly one such live id exists. Everything else stays
-/// fail-closed: no base match, or an ambiguous one, refuses the start exactly
-/// as before.
-pub(in crate::live::sessions::actor) fn resolve_requested_model_id(
-    live_ids: &[String],
-    requested: &str,
-) -> Option<String> {
-    if live_ids.iter().any(|value| value == requested) {
-        return Some(requested.to_string());
-    }
-    let requested_base = base_model_id(requested);
-    if requested_base.is_empty() {
-        return None;
-    }
-    let mut candidates = live_ids
-        .iter()
-        .filter(|value| base_model_id(value) == requested_base);
-    let first = candidates.next()?;
-    if candidates.next().is_some() {
-        return None;
-    }
-    Some(first.clone())
-}
-
-/// The id with any trailing bracket variant marker removed:
-/// `claude-fable-5[1m]` -> `claude-fable-5`.
-fn base_model_id(id: &str) -> &str {
-    match id.find('[') {
-        Some(index) => &id[..index],
-        None => id,
-    }
 }
 
 fn live_model_ids(startup_state: &SessionStartupState) -> Vec<String> {
@@ -543,101 +437,5 @@ impl SessionActor {
         }
 
         Ok(ConfigApplyState::Queued)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{absent_control_is_posture, resolve_requested_model_id};
-
-    fn live(ids: &[&str]) -> Vec<String> {
-        ids.iter().map(|value| value.to_string()).collect()
-    }
-
-    #[test]
-    fn exact_match_wins() {
-        let ids = live(&["default", "claude-fable-5", "sonnet"]);
-        assert_eq!(
-            resolve_requested_model_id(&ids, "claude-fable-5").as_deref(),
-            Some("claude-fable-5")
-        );
-    }
-
-    #[test]
-    fn rotated_variant_resolves_from_base_request() {
-        let ids = live(&["default", "opus[1m]", "claude-fable-5[1m]", "sonnet", "haiku"]);
-        assert_eq!(
-            resolve_requested_model_id(&ids, "claude-fable-5").as_deref(),
-            Some("claude-fable-5[1m]")
-        );
-    }
-
-    #[test]
-    fn variant_request_resolves_to_plain_base() {
-        let ids = live(&["default", "claude-fable-5", "sonnet"]);
-        assert_eq!(
-            resolve_requested_model_id(&ids, "claude-fable-5[1m]").as_deref(),
-            Some("claude-fable-5")
-        );
-    }
-
-    #[test]
-    fn exact_match_preferred_over_base_candidates() {
-        let ids = live(&["claude-fable-5", "claude-fable-5[1m]"]);
-        assert_eq!(
-            resolve_requested_model_id(&ids, "claude-fable-5[1m]").as_deref(),
-            Some("claude-fable-5[1m]")
-        );
-    }
-
-    #[test]
-    fn ambiguous_base_match_fails_closed() {
-        let ids = live(&["claude-fable-5[1m]", "claude-fable-5[200k]"]);
-        assert_eq!(resolve_requested_model_id(&ids, "claude-fable-5"), None);
-    }
-
-    #[test]
-    fn unknown_model_fails_closed() {
-        let ids = live(&["default", "sonnet", "haiku"]);
-        assert_eq!(resolve_requested_model_id(&ids, "claude-fable-5"), None);
-    }
-
-    #[test]
-    fn bracket_only_request_fails_closed() {
-        let ids = live(&["default", "sonnet"]);
-        assert_eq!(resolve_requested_model_id(&ids, "[1m]"), None);
-    }
-
-    // Absent-control disposition. claude narrows the control SET per model:
-    // `fast` exists only under opus, and haiku drops `effort` too. A quality
-    // control the applied model does not surface must launch without it
-    // instead of failing the start; a posture control must still refuse.
-
-    #[test]
-    fn absent_fast_control_is_not_posture() {
-        assert!(!absent_control_is_posture("fast"));
-        assert!(!absent_control_is_posture("fast_mode"));
-    }
-
-    #[test]
-    fn absent_quality_controls_are_not_posture() {
-        assert!(!absent_control_is_posture("effort"));
-        assert!(!absent_control_is_posture("reasoning_effort"));
-        assert!(!absent_control_is_posture("verbosity"));
-    }
-
-    #[test]
-    fn absent_mode_family_stays_posture() {
-        assert!(absent_control_is_posture("mode"));
-        assert!(absent_control_is_posture("collaboration_mode"));
-        assert!(absent_control_is_posture("sandbox_mode"));
-        assert!(absent_control_is_posture("approval_policy"));
-    }
-
-    #[test]
-    fn absent_model_control_never_reads_as_posture_mode() {
-        // "model" contains "mode": the model row must not be judged against
-        // the posture vocabulary.
-        assert!(!absent_control_is_posture("model"));
     }
 }
