+++ anyharness/crates/anyharness-lib/src/live/sessions/actor/config/mod.rs
@@ -1,3 +1,4 @@
+pub mod admission;
 pub mod apply;
 pub mod confirmation;
 pub mod diagnostics;
